requirements.txt
django
pillow
django-ckeditor==6.7.1
whitenoise